Reimplentation of dsiMode varible...
Now defined the same way as hbmenu. (I had forgotten about the
load_crt0.s file)

dsiMode set to 1 in load_crt0.s. If I don't do this hbemnu 0.6 booted
from this hangs. Probably because it misdetected as NTR mode.

This forces dsiMode to always be 1. The same as it was in my previous
commits, just impleneted in the same way original hbmenu would have had
it.