Added FAT12/16 support.
Also cleaned up a couple warnings. There is duplicate code in
4swordshax.s and an unused "reg" in arm9clear.arm.c

Code for drawing color on screen for debug doesn't play nice with this
(causes a jump to empty vram crash). So it has been disabled in
minitwlpayload. Tested on both FAT32 and FAT16 on hardware. It works
fine. :D

It was the DLDI code that bloats the payload. minitwlpayload did not
have this so re-enabling fat16 did not cause issues with the size of the
payload.